+++ bin/hnc/src/arguments/base.rs
@@ -88,6 +88,10 @@ pub struct HuffArgs {
     /// Print version
     #[arg(short = 'V', long = "version")]
     pub version_long: bool,
+
+    /// Output the flattened source code
+    #[clap(long = "flattened-source")]
+    pub flattened_source: bool,
 }
 
 #[derive(Subcommand, Clone, Debug)]